garage/requirements.txt
diff-match-patch==20121119
Django==1.11
django-appconf==1.0.2
django-avatar==4.1.0
django-bootstrap3==9.1.0
django-crispy--forms==1.7.2
django-filter==1.1.0
django-import-export==1.0.0
django-multiselectfield==0.1.8
django-widget-tweaks==1.4.2
et-xmlfile==1.0.1
html5lib==1.0b10
httplib2==0.11.3
jdcal==1.4
numpy==1.14.2
odfpy==1.3.6
openpyxl==2.5.3
pandas==0.22.0
Pillow==5.0.0
pinax-badges==2.0.0
psycopg2==2.7.4
py-gravatar==0.0.3
PyPDF2==1.26.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
reportlab==3.4.0
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
tablib==0.12.1
unicodecsv==0.14.1
webencodings==0.5.1
xhtml2pdf==0.2.1
xlrd==1.1.0
xlwt==1.3.0
